Moved Single Item to seperate component
