+++ inventory-app/public/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
    
-    <title>React App</title>
+    <title>Inventory App</title>
   </head>
   <body>
 <script type="text/javascript"  src="../src/index.js" defer  ></script>
