+++ inventory-app/src/App.js
@@ -4,7 +4,7 @@ import './App.css';
 import {ItemList } from './ItemList';
 import  NavBar  from './NavBar';
 import Form from './Form'
-
+import SingleItem from './SingleItem';
 
 
 function App() {
@@ -12,24 +12,13 @@ function App() {
 	const [items , setItems  ] = useState ([])
 	const [view, setView] = useState(0)
 
-	function updateForm() {
-	setView(2)
-}
+
 	const views = [
 	
  <ItemList items={items} selectedItem={selectedItem} setSelectedItem={setSelectedItem} setView={setView} />
 
 	,
-	//Single page 
-	//   <p>{JSON.stringify(selectedItem)}</p>,
-	<div className='singleItem-card'>
-			<h1 className='title'>{selectedItem.title}</h1>
-			<img src={selectedItem.image} alt=''/>
-			<p>{selectedItem.description}</p>
-			<h2 className='title'>£{selectedItem.price}</h2>
-		<button className="singleView-btn">Delete Item</button>
-		<button className="singleView-btn"onClick={updateForm}>Update Item</button>
-	</div>
+	<SingleItem items={items} selectedItem={selectedItem} setSelectedItem={setSelectedItem} setView={setView}/>
 	 ,
 	//  Form page
 	  <div><Form /></div> ,
